pyproject.toml
[tool.poetry]
name = "marvel_api_client"
version = "0.1.0"
description = "get marvel charaacter descriptions"
authors = ["Yixiao Zhou"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.5.2"
requests = "^2.28.1"
seaborn = "^0.12.1"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
myst-nb = "^0.17.1"
sphinx-autoapi = "^2.0.0"
sphinx-rtd-theme = "^1.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
